backend/requirement.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
numpy==1.24.3
openai==1.3.0